#61 - ControlMeasureCreator removed, just just CMBuilder.builder().build: ControlMeasure. Readme updated, tests, too.
